IDVA5-2053 - Code changes after code review
